requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.3
xlsxwriter==3.2.0
rich==13.9.3
pydantic==2.5.2
python-dotenv==1.0.0
loguru==0.7.2